+++ src/main/kotlin/model/Edge.kt
@@ -1,4 +1,4 @@
-package models
+package model
 
 interface Edge <E, V> {
     var element: E
