+++ src/main/kotlin/model/Vertex.kt
@@ -1,4 +1,4 @@
-package models
+package model
 
 interface Vertex<V> {
     var element: V
